+++ ios/BTE/API/Model/Exposure.swift
@@ -1,11 +1,3 @@
-//
-//  Exposure.swift
-//  BTE
-//
-//  Created by Matthew Buckley on 6/5/20.
-//  Copyright © 2020 Path Check Inc. All rights reserved.
-//
-
 import ExposureNotification
 import Foundation
