+++ ios/BTE/API/Model/ExposureConfiguration.swift
@@ -1,15 +1,8 @@
-//
-//  ExposureConfiguration.swift
-//  BTE
-//
-//  Created by Matthew Buckley on 6/5/20.
-//  Copyright © 2020 Path Check Inc. All rights reserved.
-//
-
 import Foundation
 import ExposureNotification
 
 struct ExposureConfiguration: Codable {
+
   let minimumRiskScore: ENRiskScore
   let attenuationLevelValues: [ENRiskLevelValue]
   let attenuationWeight: Double
@@ -19,4 +12,23 @@ struct ExposureConfiguration: Codable {
   let durationWeight: Double
   let transmissionRiskLevelValues: [ENRiskLevelValue]
   let transmissionRiskWeight: Double
+
+}
+
+extension ExposureConfiguration {
+
+  var asENExposureConfiguration: ENExposureConfiguration {
+    let exposureConfiguration = ENExposureConfiguration()
+    exposureConfiguration.minimumRiskScore = minimumRiskScore
+    exposureConfiguration.attenuationLevelValues = attenuationLevelValues as [NSNumber]
+    exposureConfiguration.attenuationWeight = attenuationWeight
+    exposureConfiguration.daysSinceLastExposureLevelValues = daysSinceLastExposureLevelValues as [NSNumber]
+    exposureConfiguration.daysSinceLastExposureWeight = daysSinceLastExposureWeight
+    exposureConfiguration.durationLevelValues = durationLevelValues as [NSNumber]
+    exposureConfiguration.durationWeight = durationWeight
+    exposureConfiguration.transmissionRiskLevelValues = transmissionRiskLevelValues as [NSNumber]
+    exposureConfiguration.transmissionRiskWeight = transmissionRiskWeight
+    return exposureConfiguration
+  }
+
 }
