+++ ios/BTE/API/Model/TestResult.swift
@@ -1,11 +1,3 @@
-//
-//  TestResult.swift
-//  BTE
-//
-//  Created by Matthew Buckley on 6/5/20.
-//  Copyright © 2020 Path Check Inc. All rights reserved.
-//
-
 import Foundation
 
 struct TestResult: Codable {
