+++ ios/BTE/API/Requests/DiagnosisKeyRequests.swift
@@ -1,11 +1,3 @@
-//
-//  DiagnosisKeyRequests.swift
-//  BTE
-//
-//  Created by Matthew Buckley on 6/5/20.
-//  Copyright © 2020 Path Check Inc. All rights reserved.
-//
-
 import Alamofire
 import ExposureNotification
 
@@ -32,13 +24,6 @@ enum DiagnosisKeyListRequest: APIRequest {
   var parameters: Parameters? {
     switch self {
     case .post(let diagnosisKeys):
-      // Convert keys to something that can be encoded to JSON and upload them.
-      let diagnosisKeys = diagnosisKeys.compactMap { diagnosisKey -> ExposureKey? in
-        return ExposureKey(keyData: diagnosisKey.keyData,
-                                   rollingPeriod: diagnosisKey.rollingPeriod,
-                                   rollingStartNumber: diagnosisKey.rollingStartNumber,
-                                   transmissionRiskLevel: diagnosisKey.transmissionRiskLevel)
-      }
       let params = diagnosisKeys.map { try? $0.toJson() as? JSONObject }
       return [
         "diagnosisKeys": params
