+++ ios/BTE/API/Requests/ExposureConfigurationRequests.swift
@@ -1,11 +1,3 @@
-//
-//  ExposureConfigurationRequests.swift
-//  BTE
-//
-//  Created by Matthew Buckley on 6/5/20.
-//  Copyright © 2020 Path Check Inc. All rights reserved.
-//
-
 import Alamofire
 import Foundation
