+++ ios/BTE/Extensions/ENTemporaryExposureKey+Extensions.swift
@@ -1,0 +1,11 @@
+import Foundation
+import ExposureNotification
+
+extension ENTemporaryExposureKey {
+  var asCodableKey: ExposureKey {
+    ExposureKey(keyData: keyData,
+                rollingPeriod: rollingPeriod,
+                rollingStartNumber: rollingStartNumber,
+                transmissionRiskLevel: transmissionRiskLevel)
+  }
+}
