+++ ios/BTE/String+Extensions.swift
@@ -1,11 +1,3 @@
-//
-//  String+Extensions.swift
-//  BTE
-//
-//  Created by Matthew Buckley on 6/3/20.
-//  Copyright © 2020 Path Check Inc. All rights reserved.
-//
-
 import Foundation
 
 extension String {
